Fixed import path after rename of nanny_task list, fixed import links for new db_gateways location after merge
